requirements.txt
apache-airflow-providers-apache-kafka
kafka-python
pyspark
pyhive
celery
redis
requests
pandas
pyarrow
fastparquet
great_expectations
sqlalchemy
psycopg2-binary
nltk
vaderSentiment